+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - support rendering in a repeated MapView
 - animate internally so that deck.gl _animate is not necessary anymore
 - add imageUnscale param to allow arbitrary scaled data
+- increase particle drop when zooming out
 
 ## 1.0.9